@@ -16,7 +16,7 @@ import Foundation
 /// Kind of semantic tokens which will wrap a combination of `SizingSemanticToken` depending to viewports / size classes.
 /// Kind of composite token with multiple values, but not named "composite" because this word is already used in the design system.
 /// Allows to gather the multiple-value tokens from Figma inside one object.
-public final class MultipleSizingTokens: NSObject {
+public final class MultipleSizingTokens: NSObject, Sendable {
 
     /// For **extra-compact** and **compact** viewports
     public let compact: SizingSemanticToken
