@@ -17,7 +17,7 @@ import OUDSTokensRaw
 /// Kind of semantic tokens which will wrap a combination of `TypographyCompositeRawToken` depending to size classes.
 /// Kind of composite token with multiple values, but not named "composite" because this word is already used in the design system.
 /// Allows to gather the multiple-value tokens from Figma inside one object.
-public final class MultipleTypographyTokens: NSObject {
+public final class MultipleTypographyTokens: NSObject, Sendable {
 
     /// For **extra-compact** and **compact** viewports
     public let compact: TypographyCompositeRawToken
